TS doesn't include .d.ts files on builds
My solution will include the missing files, but unfortunatelly it isn't future proof.

I tried setting up a recursive search so deeper matching files would be included, but in the copy proccess it would need to preserve the parent root, but exclude the src folder without cd'ing into it because it would require to un cd'ing into it and mess up the build proccess...

In other words, it's a problem that future me might need to deal with.

I hope not.