Added missing export
I kind of love/hate this decision.

On the one hand, the library is more open and flexible, and makes the documentation more legible.

On the other hand, it will more likely bring issues if a non TS nor JSCheck user tries to construct it, specially in the transition from v1 to v2.

I guess I will hope the users read the documentation.